requirements.txt
requests==2.32.3
python-dotenv==1.0.1
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0
